manage.py

+++ core/maker_lib/target_excel_file_maker.py
@@ -11,7 +11,7 @@ from core.excel_maker import ExcelMaker
 from core.similarity import Substitution, save_unrecognizable, read_substitution_list, get_nlp
 
 
-def make_target_excel_file(temp_file_path: str, target_excel_file_path: str):
+def make_target_excel_file(temp_file_path: str, target_excel_file_path: str, temp_remove=True):
 
     if is_open_file(target_excel_file_path):
         exit()
@@ -20,8 +20,14 @@ def make_target_excel_file(temp_file_path: str, target_excel_file_path: str):
 
     # Читаем временный файл с распарсенными данными
     df_parsed = pd.read_csv(temp_file_path)
+
     # Удаляем временный файл
-    os.remove(temp_file_path)
+    if temp_remove:
+        os.remove(temp_file_path)
+    else:
+        print(Fore.YELLOW + Style.BRIGHT + f"Временный файл не был удален так как {temp_remove=}", end='')
+        print(Style.RESET_ALL)
+
     df_parsed.dropna(subset=[stt.DRUG_COLUMN, stt.RELEASE_FORM_COLUMN], inplace=True)
 
     # Загружаем лингвистическую модель
@@ -39,15 +45,13 @@ def make_target_excel_file(temp_file_path: str, target_excel_file_path: str):
     maker = ExcelMaker(nlp=nlp, substitution=substitution)
     # Загружаем данные схем лечения
     print("Загрузка файлов с схемой лечения...")
-    maker.load(stt.TASK_DIR, columns={"#": stt.ORDER_COLUMN,
-                                      "Симптом": stt.SYMPTOM_COLUMN,
-                                      "Препарат": stt.DRUG_COLUMN,
-                                      "Форма выпуска": stt.RELEASE_FORM_COLUMN})
+    maker.load(stt.TASK_DIR, columns=stt.SCHEME_COLUMN_MAPPER)
 
     # Проверяем наличие не распознанных выражений форм выпуска препаратов
     unrecognizable_list = maker.check_model(df_parsed)
 
     if unrecognizable_list:
+        # Создается или обновляется файл для корректировки выражений форм выпуска препаратов
         save_unrecognizable(stt.UNRECOGNIZABLE_FILE_PATH, unrecognizable_list, nlp)
         print(f"Обнаружено {len(unrecognizable_list)} нераспознаваемых выражений:")
         pprint(unrecognizable_list)
@@ -60,6 +64,7 @@ def make_target_excel_file(temp_file_path: str, target_excel_file_path: str):
         # Проверяем результат
         if len(dfs) == 0:
             print(Fore.RED + Style.BRIGHT + "Не удалось найти ни один подходящий препарат.", end='')
+            print(Fore.YELLOW + Style.BRIGHT + "\n...terminated")
             print(Style.RESET_ALL)
             exit()
 
@@ -68,11 +73,13 @@ def make_target_excel_file(temp_file_path: str, target_excel_file_path: str):
                      dfs,
                      highlighted_columns=[stt.ORDER_COLUMN, stt.SYMPTOM_COLUMN,
                                           'Спосіб застосування та дози', stt.URL_COLUMN])
-        print(f"Excel файл создан и сохранен '{target_excel_file_path}'")
+
+        print("\n" + Fore.GREEN + Style.BRIGHT + f"Excel файл создан и сохранен: '{target_excel_file_path}'", end='')
+        print(Style.RESET_ALL)
 
         log_path = os.path.join(stt.LOG_DIR, stt.DomainKeys.LIKITEKA.name.lower(), 'release-form-similarity.log')
         save_log(log_path, refused)
-        print(f"Лог файл с отклоненными записями сохранен '{log_path}'")
+        print(f"Лог файл с {len(refused)} отклоненными записями сохранен: '{log_path}'")
 
         if not refused:
             print('Не обнаружено отклоненных записей')
@@ -82,5 +89,5 @@ if __name__ == "__main__":
     temp_file_path = os.path.join(stt.TEMP_DATA_DIR, stt.DomainKeys.LIKITEKA.name.lower() + "_parsed.csv")
     # Путь куда будет сохранен файл
     target_excel_file_path = os.path.join(stt.EXCEL_DATA_DIR, stt.DomainKeys.LIKITEKA.name.lower() + ".xlsx")
-    make_target_excel_file(temp_file_path, target_excel_file_path)
+    make_target_excel_file(temp_file_path, target_excel_file_path, temp_remove=False)
